ChatExample
This sample code for basic chat app allows for multiple users to connect to the same server and all talk to eachother in same chat board. Can't distinguish users from eachother, but that's why its basic ;-).  This example needs to change the ip address within the App.js file in mobile folder, to your computers local address and then it should work.